requirements.txt
flask
flask-mqtt
redis
paho-mqtt
pandas